# swagger.yaml
openapi: 3.0.0
info:
  title: Fitness API
  version: 1.0.0
  description: API voor het beheren van oefeningen, uitgevoerde oefeningen, sets, gebruikers en workouts.
paths:
...
  /workouts/{workoutId}/link/{performedExerciseId}:
    post:
      summary: Link a performed exercise to a workout
      operationId: linkPerformedExerciseToWorkout
      parameters:
        - name: workoutId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the workout to which the exercise will be linked
        - name: performedExerciseId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the performed exercise to link to the workout
      responses:
        '200':
          description: Performed Exercise successfully linked to the Workout
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Performed Exercise linked successfully.
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
  /workouts:
    get:
      summary: Get all workouts
      operationId: getWorkouts
      responses:
        '200':
          description: OK
    post:
      summary: Create a new workout
      operationId: createWorkout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkoutDTO'
      responses:
        '201':
          description: Created
  /workouts/{id}:
    get:
      summary: Get a workout by ID
      operationId: getWorkoutById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
    put:
      summary: Update a workout by ID
      operationId: updateWorkout
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkoutDTO'
      responses:
        '200':
          description: OK
    delete:
      summary: Delete a workout by ID
      operationId: deleteWorkout
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No Content
components:
  schemas:
...
    WorkoutDTO:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        duration:
          type: number